// backend/src/proto/bookStorePackage/Book.ts
// Original file: ../proto/bookStore.proto


export interface Book {
  'title'?: (string);
  'author'?: (string);
  'description'?: (string);
  'price'?: (number | string);
  'disc'?: (number);
}

export interface Book__Output {
  'title'?: (string);
  'author'?: (string);
  'description'?: (string);
  'price'?: (number);
  'disc'?: (number);
}